requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Diffusion Models
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
xformers>=0.0.20  # Optional: for memory-efficient attention

# 3D & SLAM
open3d>=0.17.0
opencv-python>=4.8.0
kornia>=0.7.0
pytorch3d>=0.7.4  # Optional: for 3D operations

# Medical Imaging
SimpleITK>=2.2.0
nibabel>=5.0.0
scikit-image>=0.21.0
monai>=1.2.0

# Training & Logging
hydra-core>=1.3.0
wandb>=0.15.0
tensorboard>=2.13.0
einops>=0.6.0
timm>=0.9.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=9.5.0
pyyaml>=6.0
albumentations>=1.3.0

# 3D Gaussian Splatting (manual install recommended)
# git+https://github.com/graphdeco-inria/gaussian-splatting.git
# or use diff-gaussian-rasterization package
